backend/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
requests
chromadb
numpy
sentence-transformers
dateparser